Update 1.0.6
Bug fixes, for example, when tolerances were low, the solver never converged
despite there being a check if previous error was the same as current error.
I added an epsilon value and compared the difference towards it, considering
convergence to have failed if the difference is lower than the epsilon value.

Added a destructor, which frees the memory of chain and chain->joints. This is
necessary when creating a Fabrik2D instance from within a function or a
routine.

Added some optional parameters to funcitons and changed pointer arguments
(int* lengths) to their array representation (int lengths[]).

Added evaluation on FABRIK2D execution time vs other approaches.